Aggiunta autenticazione tramite JWT

+++ UserService.Business/Services/UserService.cs
@@ -24,56 +24,70 @@ namespace UserService.Business.Services
         public async Task<List<UserDto>> GetAllAsync()
         {
             List<User> users = await repository.GetAllAsync();
-            List<UserDto> userDtos = mapper.Map<List<UserDto>>(users);
-            return userDtos;
+            return mapper.Map<List<UserDto>>(users);
         }
 
-        public async Task<UserDto?> GetUserByIdAsync(int id)
+        public async Task<UserDto> GetUserByIdAsync(int id)
         {
-            User? user = await repository.GetUserByIdAsync(id);
-            UserDto? userDto = mapper.Map<UserDto>(user);
-            return userDto;
+            User user = await repository.GetUserByIdAsync(id) ?? throw new KeyNotFoundException($"Utente con ID {id} non trovato.");
+
+            return mapper.Map<UserDto>(user);
         }
-        public async Task AddAsync(CreateUserDto user)
+
+        public async Task AddAsync(CreateUserDto userDto)
         {
-            User u = mapper.Map<User>(user);
+            User? existingUser = await repository.GetUserByUsernameAsync(userDto.UserName);
+            if (existingUser != null)
+                throw new InvalidOperationException($"Lo username '{userDto.UserName}' è già in uso.");
+
+            User u = mapper.Map<User>(userDto);
+
             u.LastModified = DateTime.UtcNow;
-            u.Password = BCrypt.Net.BCrypt.HashPassword(user.Password);
+            u.Password = BCrypt.Net.BCrypt.HashPassword(userDto.Password);
+
             await repository.AddAsync(u);
         }
 
-        public async Task DeleteAsync(int id)
+        public async Task DeleteAsync(int id, int userId)
         {
+            if (userId != id) throw new UnauthorizedAccessException("Non hai i permessi per eliminare questo utente.");
+            User user = await repository.GetUserByIdAsync(id) ?? throw new KeyNotFoundException($"Impossibile eliminare: utente con ID {id} non trovato.");
             await repository.DeleteAsync(id);
         }
 
-        public async Task UpdateAsync(int id, UpdateUserDto user)
+        public async Task UpdateAsync(int id, UpdateUserDto userDto, int userId)
         {
-            User? userEntity = await repository.GetUserByIdAsync(id);
-            if (userEntity == null)
-                throw new Exception("Utente non trovato");
-            mapper.Map(user, userEntity);
-            await repository.UpdateAsync(userEntity);
+            if (userId != id) throw new UnauthorizedAccessException("Non hai i permessi per modificare questo utente.");
+            User user = await repository.GetUserByIdAsync(id) ?? throw new KeyNotFoundException($"Impossibile eliminare: utente con ID {id} non trovato.");
+
+            mapper.Map(userDto, user);
+            user.LastModified = DateTime.UtcNow;
+
+            await repository.UpdateAsync(user);
         }
 
-        public async Task<bool> ChangePasswordAsync(int id, ChangePasswordDto dto)
+        public async Task ChangePasswordAsync(int id, ChangePasswordDto dto, int userId)
         {
-            User? userEntity = await repository.GetUserByIdAsync(id);
-            if (userEntity == null)
-                throw new Exception("Utente non trovato");
-            bool control = BCrypt.Net.BCrypt.Verify(dto.OldPassword, userEntity.Password);
-            if(!control)
-                return false;
-            userEntity.Password = BCrypt.Net.BCrypt.HashPassword(dto.NewPassword);
-            await repository.UpdateAsync(userEntity);
-            return true;
+            if (userId != id) throw new UnauthorizedAccessException("Non hai i permessi per modificare questo utente.");
+            User user = await repository.GetUserByIdAsync(id) ?? throw new KeyNotFoundException($"Impossibile eliminare: utente con ID {id} non trovato.");
+
+            bool isOldPasswordCorrect = BCrypt.Net.BCrypt.Verify(dto.OldPassword, user.Password);
+            if (!isOldPasswordCorrect) throw new UnauthorizedAccessException("La vecchia password non è corretta.");
+
+            if (dto.OldPassword == dto.NewPassword) throw new InvalidOperationException("La nuova password non può essere uguale alla precedente.");
+
+            user.Password = BCrypt.Net.BCrypt.HashPassword(dto.NewPassword);
+            user.LastModified = DateTime.UtcNow;
+
+            await repository.UpdateAsync(user);
         }
-        public async Task<string?> LoginAsync(LoginDto dto)
+
+        public async Task<string> LoginAsync(LoginDto dto)
         {
-            User? userEntity = await repository.GetUserByUsernameAsync(dto.Username);
-            if (userEntity == null) return null;
-            bool control = BCrypt.Net.BCrypt.Verify(dto.Password, userEntity.Password);
-            if (!control) return null;
+            User userEntity = await repository.GetUserByUsernameAsync(dto.Username) ?? throw new UnauthorizedAccessException("Credenziali non valide.");
+
+            if (userEntity == null || !BCrypt.Net.BCrypt.Verify(dto.Password, userEntity.Password)) throw new UnauthorizedAccessException("Credenziali non valide.");
+
             return GenerateJwtToken(userEntity);
         }
         private string GenerateJwtToken(User user)
